#2 Move CMA-related tests into separate file and ensure all new lines covered
