+++ src/tests/dynamo0p3_cma_test.py
@@ -1,0 +1,377 @@
+# -----------------------------------------------------------------------------
+# BSD 3-Clause License
+#
+# Copyright (c) 2017, Science and Technology Facilities Council
+# (c) The copyright relating to this work is owned jointly by the Crown,
+# Met Office and NERC 2016.
+# However, it has been created with the help of the GungHo Consortium,
+# whose members are identified at https://puma.nerc.ac.uk/trac/GungHo/wiki
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#
+# * Redistributions of source code must retain the above copyright notice, this
+#   list of conditions and the following disclaimer.
+#
+# * Redistributions in binary form must reproduce the above copyright notice,
+#   this list of conditions and the following disclaimer in the documentation
+#   and/or other materials provided with the distribution.
+#
+# * Neither the name of the copyright holder nor the names of its
+#   contributors may be used to endorse or promote products derived from
+#   this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+# COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+# -----------------------------------------------------------------------------
+# Author R. Ford and A. R. Porter, STFC Daresbury Lab
+
+''' This module tests the support for Column-Matrix-Assembly operators in
+the Dynamo 0.3 API using pytest. '''
+
+import pytest
+import fparser
+from parse import ParseError
+from fparser import api as fpapi
+from dynamo0p3 import DynKernMetadata
+
+CMA_ASSEMBLE = '''
+module testkern_cma
+  type, extends(kernel_type) :: testkern_cma_type
+     type(arg_type), meta_args(3) =                 &
+          (/ arg_type(gh_operator,gh_read, any_space_1, any_space_2), &
+             arg_type(gh_columnwise_operator,gh_write, any_space_1,   &
+                      any_space_2),                                   &
+             arg_type(gh_field,gh_read, any_space_1)                  &
+           /)
+     integer, parameter :: iterates_over = cells
+   contains
+     procedure() :: code => testkern_cma_code
+  end type testkern_cma_type
+contains
+  subroutine testkern_cma_code(a,b,c,d)
+  end subroutine testkern_cma_code
+end module testkern_cma
+'''
+
+
+def test_cma_mdata_assembly():
+    ''' Check that we can parse meta-data entries relating to Column-Matrix
+    Assembly '''
+    fparser.logging.disable('CRITICAL')
+    code = CMA_ASSEMBLE
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    dkm = DynKernMetadata(ast, name=name)
+    dkm_str = str(dkm.arg_descriptors[1])
+    expected = (
+        "DynArgDescriptor03 object\n"
+        "  argument_type[0]='gh_columnwise_operator'\n"
+        "  access_descriptor[1]='gh_write'\n"
+        "  function_space_to[2]='any_space_1'\n"
+        "  function_space_from[3]='any_space_2'\n")
+    print dkm_str
+    assert expected in dkm_str
+
+
+def test_cma_mdata_assembly_missing_op():
+    ''' Check that we raise the expected error if the supplied meta-data
+    is assembling a gh_columnwise_operator but doesn't have a read-only
+    gh_operator '''
+    fparser.logging.disable('CRITICAL')
+    # Remove  the (required) LMA operator
+    code = CMA_ASSEMBLE.replace(
+        "arg_type(gh_operator,gh_read, any_space_1, any_space_2),", "", 1)
+    code = code.replace("meta_args(3) =", "meta_args(2) =", 1)
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    with pytest.raises(ParseError) as excinfo:
+        _ = DynKernMetadata(ast, name=name)
+    assert ("Kernel testkern_cma_type assembles a column-wise operator but "
+            "does not have any LMA operators as read-only arguments") in \
+        str(excinfo)
+
+
+def test_cma_mdata_multi_writes():
+    ''' Check that we raise the expected error if the supplied meta-data
+    specifies more than one CMA operator that is written to '''
+    fparser.logging.disable('CRITICAL')
+    # Replace the field arg with another CMA operator that is written to
+    code = CMA_ASSEMBLE.replace(
+        "arg_type(gh_field,gh_read, any_space_1)",
+        "arg_type(gh_columnwise_operator,gh_write,any_space_1,any_space_2)",
+        1)
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    with pytest.raises(ParseError) as excinfo:
+        _ = DynKernMetadata(ast, name=name)
+    assert ("A Dynamo 0.3 kernel cannot update more than one CMA "
+            "(column-wise) operator but kernel testkern_cma_type "
+            "updates 2") in str(excinfo)
+    code = CMA_ASSEMBLE.replace(
+        "arg_type(gh_field,gh_read, any_space_1)",
+        "arg_type(gh_columnwise_operator,gh_write,any_space_1,any_space_2),&\n"
+        "arg_type(gh_columnwise_operator,gh_write,any_space_1,any_space_2)",
+        1)
+    code = code.replace("meta_args(3) = ", "meta_args(4) = ", 1)
+    ast = fpapi.parse(code, ignore_comments=False)
+    with pytest.raises(ParseError) as excinfo:
+        _ = DynKernMetadata(ast, name=name)
+    assert ("A Dynamo 0.3 kernel cannot update more than one CMA "
+            "(column-wise) operator but kernel testkern_cma_type "
+            "updates 3") in str(excinfo)
+
+
+def test_cma_mdata_mutable_op():
+    ''' Check that we raise the expected error if the supplied meta-data
+    is assembling a gh_columnwise_operator but doesn't have a read-only
+    gh_operator '''
+    fparser.logging.disable('CRITICAL')
+    # Make the LMA operator gh_write instead of gh_read
+    code = CMA_ASSEMBLE.replace(
+        "arg_type(gh_operator,gh_read, any_space_1, any_space_2),",
+        "arg_type(gh_operator,gh_write, any_space_1, any_space_2),", 1)
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    with pytest.raises(ParseError) as excinfo:
+        _ = DynKernMetadata(ast, name=name)
+    assert ("Kernel testkern_cma_type assembles a column-wise operator but "
+            "does not have any LMA operators as read-only arguments") in \
+        str(excinfo)
+
+
+def test_cma_mdata_writes_lma_op():
+    ''' Check that we raise the expected error if the supplied meta-data
+    is assembling a gh_columnwise_operator but also writes to a
+    gh_operator '''
+    fparser.logging.disable('CRITICAL')
+    # Add an additional LMA operator that has write access
+    code = CMA_ASSEMBLE.replace(
+        "arg_type(gh_operator,gh_read, any_space_1, any_space_2), &\n",
+        "arg_type(gh_operator,gh_read, any_space_1, any_space_2), &\n"
+        "arg_type(gh_operator,gh_write, any_space_1, any_space_2), &\n", 1)
+    code = code.replace("meta_args(3)", "meta_args(4)", 1)
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    with pytest.raises(ParseError) as excinfo:
+        _ = DynKernMetadata(ast, name=name)
+    assert ("Kernel testkern_cma_type assembles a column-wise operator but "
+            "also writes to a LMA operator. This is not allowed.") in \
+        str(excinfo)
+
+
+def test_cma_mdata_assembly_wrong_spaces():
+    ''' Check that we raise the expected error if the supplied meta-data
+    is assembling a gh_columnwise_operator but the to/from spaces don't
+    match those of the supplied gh_operator '''
+    fparser.logging.disable('CRITICAL')
+    # Remove  the (required) LMA operator
+    code = CMA_ASSEMBLE.replace(
+        "arg_type(gh_operator,gh_read, any_space_1, any_space_2),",
+        "arg_type(gh_operator,gh_read, any_space_3, any_space_2),", 1)
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    with pytest.raises(ParseError) as excinfo:
+        _ = DynKernMetadata(ast, name=name)
+    assert ("When assembling a column-wise operator from LMA operators the "
+            "to and from function spaces must match") in str(excinfo)
+    code = CMA_ASSEMBLE.replace(
+        "arg_type(gh_operator,gh_read, any_space_1, any_space_2),",
+        "arg_type(gh_operator,gh_read, any_space_1, any_space_3),", 1)
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    with pytest.raises(ParseError) as excinfo:
+        _ = DynKernMetadata(ast, name=name)
+    assert ("When assembling a column-wise operator from LMA operators the "
+            "to and from function spaces must match") in str(excinfo)
+
+
+CMA_APPLY = '''
+module testkern_cma_apply
+  type, extends(kernel_type) :: testkern_cma_type
+  type(arg_type) :: meta_args(3) = (/                                      &
+       arg_type(GH_FIELD,    GH_INC,  ANY_SPACE_1),                        &
+       arg_type(GH_FIELD,    GH_READ, ANY_SPACE_2),                        &
+       arg_type(GH_COLUMNWISE_OPERATOR, GH_READ, ANY_SPACE_1, ANY_SPACE_2) &
+       /)
+     integer, parameter :: iterates_over = cells
+   contains
+     procedure() :: code => testkern_cma_code
+  end type testkern_cma_type
+contains
+  subroutine testkern_cma_code(a,b,c,d)
+  end subroutine testkern_cma_code
+end module testkern_cma_apply
+'''
+
+
+def test_cma_mdata_apply():
+    ''' Check that we can parse meta-data entries relating to the
+    application of Column-Matrix operators '''
+    fparser.logging.disable('CRITICAL')
+    code = CMA_APPLY
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    dkm = DynKernMetadata(ast, name=name)
+    dkm_str = str(dkm.arg_descriptors[1])
+    expected = (
+        "DynArgDescriptor03 object\n"
+        "  argument_type[0]='gh_field'\n"
+        "  access_descriptor[1]='gh_read'\n"
+        "  function_space[2]='any_space_2'\n")
+    print dkm_str
+    assert expected in dkm_str
+    dkm_str = str(dkm.arg_descriptors[2])
+    expected = (
+        "DynArgDescriptor03 object\n"
+        "  argument_type[0]='gh_columnwise_operator'\n"
+        "  access_descriptor[1]='gh_read'\n"
+        "  function_space_to[2]='any_space_1'\n"
+        "  function_space_from[3]='any_space_2'\n")
+    assert expected in dkm_str
+
+
+def test_cma_mdata_apply_too_many_ops():
+    ''' Check that we raise the expected error if there are too-many
+    CMA operators '''
+    fparser.logging.disable('CRITICAL')
+    # Add an additional read-only CMA operator to the argument list
+    code = CMA_APPLY.replace(
+        "ANY_SPACE_2),                        &\n",
+        "ANY_SPACE_2),                        &\n"
+        "       arg_type(GH_COLUMNWISE_OPERATOR, GH_READ, ANY_SPACE_1, "
+        "ANY_SPACE_2), &\n", 1)
+    code = code.replace("meta_args(3)", "meta_args(4)", 1)
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    with pytest.raises(ParseError) as excinfo:
+        _ = DynKernMetadata(ast, name=name)
+    assert ("a kernel that applies a CMA operator must only have one such "
+            "operator in its list of arguments but found 2") in str(excinfo)
+
+
+def test_cma_mdata_apply_too_many_flds():
+    ''' Check that we raise the expected error if there are too-many
+    field args to a kernel that applies a CMA operator '''
+    fparser.logging.disable('CRITICAL')
+    # Add an additional read-only field to the argument list
+    code = CMA_APPLY.replace(
+        "ANY_SPACE_2),                        &\n",
+        "ANY_SPACE_2),                        &\n"
+        "       arg_type(GH_FIELD,    GH_READ, ANY_SPACE_2), &\n", 1)
+    code = code.replace("meta_args(3)", "meta_args(4)", 1)
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    with pytest.raises(ParseError) as excinfo:
+        _ = DynKernMetadata(ast, name=name)
+    assert ("a kernel that applies a CMA operator must have 3 arguments "
+            "(the operator and two fields) but kernel testkern_cma_type "
+            "has 4") in str(excinfo)
+
+
+def test_cma_mdata_apply_no_read_fld():
+    ''' Check that we raise the expected error if there is no read-only
+    field arg to a kernel that applies a CMA operator '''
+    fparser.logging.disable('CRITICAL')
+    # Make the read-only field gh_write instead
+    code = CMA_APPLY.replace(
+        "arg_type(GH_FIELD,    GH_READ, ANY_SPACE_2), ",
+        "arg_type(GH_FIELD,    GH_WRITE, ANY_SPACE_2), ", 1)
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    with pytest.raises(ParseError) as excinfo:
+        _ = DynKernMetadata(ast, name=name)
+    assert ("has a read-only CMA operator. In order to apply it the kernel "
+            "must have one read-only field argument") in str(excinfo)
+
+
+def test_cma_mdata_apply_no_write_fld():
+    ''' Check that we raise the expected error if there is no written
+    field arg to a kernel that applies a CMA operator '''
+    fparser.logging.disable('CRITICAL')
+    # Turn the written field into an operator instead
+    code = CMA_APPLY.replace(
+        "arg_type(GH_FIELD,    GH_INC,  ANY_SPACE_1), ",
+        "arg_type(GH_OPERATOR, GH_WRITE, ANY_SPACE_1, ANY_SPACE_1), ", 1)
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    with pytest.raises(ParseError) as excinfo:
+        _ = DynKernMetadata(ast, name=name)
+    assert ("has a read-only CMA operator. In order to apply it the kernel "
+            "must write to one field argument") in str(excinfo)
+
+
+def test_cma_mdata_apply_wrong_spaces():
+    ''' Check that we raise the expected error if the function spaces of the
+    read and write fields do not match the from and to function spaces of the
+    CMA operator '''
+    fparser.logging.disable('CRITICAL')
+    # Change the space of the field that is written
+    code = CMA_APPLY.replace("arg_type(GH_FIELD,    GH_INC,  ANY_SPACE_1)",
+                             "arg_type(GH_FIELD,    GH_INC,  ANY_SPACE_3)", 1)
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    with pytest.raises(ParseError) as excinfo:
+        _ = DynKernMetadata(ast, name=name)
+    assert ("applies a CMA operator but the function space of the field "
+            "argument it writes to (any_space_3) does not match the 'to' "
+            "space of the operator (any_space_1)") in str(excinfo)
+    # Change the space of the field that is read
+    code = CMA_APPLY.replace("arg_type(GH_FIELD,    GH_READ, ANY_SPACE_2)",
+                             "arg_type(GH_FIELD,    GH_READ, ANY_SPACE_3)", 1)
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    with pytest.raises(ParseError) as excinfo:
+        _ = DynKernMetadata(ast, name=name)
+    assert ("applies a CMA operator but the function space of the field "
+            "argument it reads from (any_space_3) does not match the 'from' "
+            "space of the operator (any_space_2)") in str(excinfo)
+
+
+CMA_MATRIX = '''
+module testkern_cma_matrix_matrix
+  type, extends(kernel_type) :: testkern_cma_type
+  type(arg_type) :: meta_args(3) = (/                                      &
+       arg_type(GH_COLUMNWISE_OPERATOR, GH_READ, ANY_SPACE_1, ANY_SPACE_2),&
+       arg_type(GH_COLUMNWISE_OPERATOR, GH_READ, ANY_SPACE_1, ANY_SPACE_2),&
+       arg_type(GH_COLUMNWISE_OPERATOR, GH_WRITE,ANY_SPACE_1, ANY_SPACE_2) &
+       /)
+     integer, parameter :: iterates_over = cells
+   contains
+     procedure() :: code => testkern_cma_code
+  end type testkern_cma_type
+contains
+  subroutine testkern_cma_code(a,b,c,d)
+  end subroutine testkern_cma_code
+end module testkern_cma_matrix_matrix
+'''
+
+
+def test_cma_mdata_matrix_prod():
+    ''' Check that we can parse meta-data entries relating to a kernel
+    that performs a product of two CMA operators '''
+    fparser.logging.disable('CRITICAL')
+    code = CMA_MATRIX
+    ast = fpapi.parse(code, ignore_comments=False)
+    name = "testkern_cma_type"
+    dkm = DynKernMetadata(ast, name=name)
+    dkm_str = str(dkm.arg_descriptors[1])
+    expected = (
+        "DynArgDescriptor03 object\n"
+        "  argument_type[0]='gh_columnwise_operator'\n"
+        "  access_descriptor[1]='gh_read'\n"
+        "  function_space_to[2]='any_space_1'\n"
+        "  function_space_from[3]='any_space_2'\n")
+    print dkm_str
+    assert expected in dkm_str
