+++ src/tests/dynamo0p3_test.py
@@ -6373,260 +6373,3 @@ def test_kernstubarglist_arglist_error():
         "Internal error. The argument list in KernStubArgList:arglist() is "
         "empty. Has the generate() method been "
         "called?") in str(excinfo.value)
-
-
-CMA_ASSEMBLE = '''
-module testkern_cma
-  type, extends(kernel_type) :: testkern_cma_type
-     type(arg_type), meta_args(3) =                 &
-          (/ arg_type(gh_operator,gh_read, any_space_1, any_space_2), &
-             arg_type(gh_columnwise_operator,gh_write, any_space_1,   &
-                      any_space_2),                                   &
-             arg_type(gh_field,gh_read, any_space_1)                  &
-           /)
-     type(func_type), dimension(2) :: meta_funcs =  &
-          (/ func_type(any_space_1, gh_column_banded_dofmap), &
-             func_type(any_space_2, gh_column_banded_dofmap) &
-           /)
-     integer, parameter :: iterates_over = cells
-   contains
-     procedure() :: code => testkern_cma_code
-  end type testkern_cma_type
-contains
-  subroutine testkern_cma_code(a,b,c,d)
-  end subroutine testkern_cma_code
-end module testkern_cma
-'''
-
-
-def test_cma_mdata_assembly():
-    ''' Check that we can parse meta-data entries relating to Column-Matrix
-    Assembly '''
-    fparser.logging.disable('CRITICAL')
-    code = CMA_ASSEMBLE
-    ast = fpapi.parse(code, ignore_comments=False)
-    name = "testkern_cma_type"
-    dkm = DynKernMetadata(ast, name=name)
-    dkm_str = str(dkm.arg_descriptors[1])
-    expected = (
-        "DynArgDescriptor03 object\n"
-        "  argument_type[0]='gh_columnwise_operator'\n"
-        "  access_descriptor[1]='gh_write'\n"
-        "  function_space_to[2]='any_space_1'\n"
-        "  function_space_from[3]='any_space_2'\n")
-    print dkm_str
-    assert expected in dkm_str
-
-
-@pytest.mark.xfail(
-    reason="Yet to decide whether metadata should include dofmap")
-def test_cma_mdata_assembly_missing_dmap():
-    ''' Check that we raise the expected error when the meta-data for a
-    CMA assembly kernel does not specify the necessary dofmap '''
-    fparser.logging.disable('CRITICAL')
-    # Remove one of the (required) column-banded dofmaps
-    code = CMA_ASSEMBLE.replace(
-        "func_type(any_space_1, gh_column_banded_dofmap),", "", 1)
-    code = code.replace("dimension(2) :: meta_funcs",
-                        "dimension(1) :: meta_funcs", 1)
-    ast = fpapi.parse(code, ignore_comments=False)
-    name = "testkern_cma_type"
-    with pytest.raises(ParseError) as excinfo:
-        _ = DynKernMetadata(ast, name=name)
-    assert ("meta-data must contain a gh_column_banded_dofmap for every "
-            "function space (whether to or from) associated with a "
-            "gh_columnwise_operator but did not find one for space "
-            "any_space_1") in str(excinfo)
-
-
-def test_cma_mdata_assembly_missing_op():
-    ''' Check that we raise the expected error if the supplied meta-data
-    is assembling a gh_columnwise_operator but doesn't have a read-only
-    gh_operator '''
-    fparser.logging.disable('CRITICAL')
-    # Remove  the (required) LMA operator
-    code = CMA_ASSEMBLE.replace(
-        "arg_type(gh_operator,gh_read, any_space_1, any_space_2),", "", 1)
-    code = code.replace("meta_args(3) =", "meta_args(2) =", 1)
-    ast = fpapi.parse(code, ignore_comments=False)
-    name = "testkern_cma_type"
-    with pytest.raises(ParseError) as excinfo:
-        _ = DynKernMetadata(ast, name=name)
-    assert ("Kernel testkern_cma_type assembles a column-wise operator but "
-            "does not have any LMA operators as read-only arguments") in \
-        str(excinfo)
-
-
-def test_cma_mdata_multi_writes():
-    ''' Check that we raise the expected error if the supplied meta-data
-    specifies more than one CMA operator that is written to '''
-    fparser.logging.disable('CRITICAL')
-    # Replace the field arg with another CMA operator that is written to
-    code = CMA_ASSEMBLE.replace(
-        "arg_type(gh_field,gh_read, any_space_1)",
-        "arg_type(gh_columnwise_operator,gh_write,any_space_1,any_space_2)",
-        1)
-    ast = fpapi.parse(code, ignore_comments=False)
-    name = "testkern_cma_type"
-    with pytest.raises(ParseError) as excinfo:
-        _ = DynKernMetadata(ast, name=name)
-    assert ("A Dynamo 0.3 kernel cannot update more than one CMA "
-            "(column-wise) operator but kernel testkern_cma_type "
-            "updates 2") in str(excinfo)
-    code = CMA_ASSEMBLE.replace(
-        "arg_type(gh_field,gh_read, any_space_1)",
-        "arg_type(gh_columnwise_operator,gh_write,any_space_1,any_space_2),&\n"
-        "arg_type(gh_columnwise_operator,gh_write,any_space_1,any_space_2)",
-        1)
-    code = code.replace("meta_args(3) = ", "meta_args(4) = ", 1)
-    ast = fpapi.parse(code, ignore_comments=False)
-    with pytest.raises(ParseError) as excinfo:
-        _ = DynKernMetadata(ast, name=name)
-    assert ("A Dynamo 0.3 kernel cannot update more than one CMA "
-            "(column-wise) operator but kernel testkern_cma_type "
-            "updates 3") in str(excinfo)
-
-
-def test_cma_mdata_mutable_op():
-    ''' Check that we raise the expected error if the supplied meta-data
-    is assembling a gh_columnwise_operator but doesn't have a read-only
-    gh_operator '''
-    fparser.logging.disable('CRITICAL')
-    # Make the LMA operator gh_write instead of gh_read
-    code = CMA_ASSEMBLE.replace(
-        "arg_type(gh_operator,gh_read, any_space_1, any_space_2),",
-        "arg_type(gh_operator,gh_write, any_space_1, any_space_2),", 1)
-    ast = fpapi.parse(code, ignore_comments=False)
-    name = "testkern_cma_type"
-    with pytest.raises(ParseError) as excinfo:
-        _ = DynKernMetadata(ast, name=name)
-    assert ("Kernel testkern_cma_type assembles a column-wise operator but "
-            "does not have any LMA operators as read-only arguments") in \
-        str(excinfo)
-
-
-def test_cma_mdata_assembly_wrong_spaces():
-    ''' Check that we raise the expected error if the supplied meta-data
-    is assembling a gh_columnwise_operator but the to/from spaces don't
-    match those of the supplied gh_operator '''
-    fparser.logging.disable('CRITICAL')
-    # Remove  the (required) LMA operator
-    code = CMA_ASSEMBLE.replace(
-        "arg_type(gh_operator,gh_read, any_space_1, any_space_2),",
-        "arg_type(gh_operator,gh_read, any_space_3, any_space_2),", 1)
-    ast = fpapi.parse(code, ignore_comments=False)
-    name = "testkern_cma_type"
-    with pytest.raises(ParseError) as excinfo:
-        _ = DynKernMetadata(ast, name=name)
-    assert ("When assembling a column-wise operator from LMA operators the "
-            "to and from function spaces must match") in str(excinfo)
-    code = CMA_ASSEMBLE.replace(
-        "arg_type(gh_operator,gh_read, any_space_1, any_space_2),",
-        "arg_type(gh_operator,gh_read, any_space_1, any_space_3),", 1)
-    ast = fpapi.parse(code, ignore_comments=False)
-    name = "testkern_cma_type"
-    with pytest.raises(ParseError) as excinfo:
-        _ = DynKernMetadata(ast, name=name)
-    assert ("When assembling a column-wise operator from LMA operators the "
-            "to and from function spaces must match") in str(excinfo)
-
-
-CMA_APPLY = '''
-module testkern_cma_apply
-  type, extends(kernel_type) :: testkern_cma_type
-  type(arg_type) :: meta_args(3) = (/                                      &
-       arg_type(GH_FIELD,    GH_INC,  ANY_SPACE_1),                        &
-       arg_type(GH_FIELD,    GH_READ, ANY_SPACE_2),                        &
-       arg_type(GH_COLUMNWISE_OPERATOR, GH_READ, ANY_SPACE_1, ANY_SPACE_2) &
-       /)
-  type(func_type) :: meta_funcs(2) = (/                                    &
-       func_type(ANY_SPACE_1, GH_COLUMN_INDIRECTION_DOFMAP),               &
-       func_type(ANY_SPACE_2, GH_COLUMN_INDIRECTION_DOFMAP)                &
-       /)
-     integer, parameter :: iterates_over = cells
-   contains
-     procedure() :: code => testkern_cma_code
-  end type testkern_cma_type
-contains
-  subroutine testkern_cma_code(a,b,c,d)
-  end subroutine testkern_cma_code
-end module testkern_cma_apply
-'''
-
-
-@pytest.mark.xfail(
-    reason="Yet to decide whether metadata should include dofmap")
-def test_cma_mdata_apply():
-    ''' Check that we can parse meta-data entries relating to the
-    application of Column-Matrix operators '''
-    fparser.logging.disable('CRITICAL')
-    code = CMA_APPLY
-    ast = fpapi.parse(code, ignore_comments=False)
-    name = "testkern_cma_type"
-    dkm = DynKernMetadata(ast, name=name)
-    dkm_str = str(dkm.arg_descriptors[1])
-    expected = (
-        "DynArgDescriptor03 object\n"
-        "  argument_type[0]='gh_columnwise_operator'\n"
-        "  access_descriptor[1]='gh_read'\n"
-        "  function_space_to[2]='any_space_1'\n"
-        "  function_space_from[3]='any_space_2'\n")
-    print dkm_str
-    assert expected in dkm_str
-
-
-CMA_MATRIX = '''
-module testkern_cma_apply
-  type, extends(kernel_type) :: testkern_cma_type
-  type(arg_type) :: meta_args(3) = (/                                      &
-       arg_type(GH_COLUMNWISE_OPERATOR, GH_READ, ANY_SPACE_1, ANY_SPACE_2),&
-       arg_type(GH_COLUMNWISE_OPERATOR, GH_READ, ANY_SPACE_1, ANY_SPACE_2),&
-       arg_type(GH_COLUMNWISE_OPERATOR, GH_WRITE,ANY_SPACE_1, ANY_SPACE_2) &
-       /)
-     integer, parameter :: iterates_over = cells
-   contains
-     procedure() :: code => testkern_cma_code
-  end type testkern_cma_type
-contains
-  subroutine testkern_cma_code(a,b,c,d)
-  end subroutine testkern_cma_code
-end module testkern_cma_apply
-'''
-
-
-def test_cma_mdata_matrix_prod():
-    ''' Check that we can parse meta-data entries relating to a kernel
-    that performs a product of two CMA operators '''
-    fparser.logging.disable('CRITICAL')
-    code = CMA_MATRIX
-    ast = fpapi.parse(code, ignore_comments=False)
-    name = "testkern_cma_type"
-    dkm = DynKernMetadata(ast, name=name)
-    dkm_str = str(dkm.arg_descriptors[1])
-    expected = (
-        "DynArgDescriptor03 object\n"
-        "  argument_type[0]='gh_columnwise_operator'\n"
-        "  access_descriptor[1]='gh_read'\n"
-        "  function_space_to[2]='any_space_1'\n"
-        "  function_space_from[3]='any_space_2'\n")
-    print dkm_str
-    assert expected in dkm_str
-
-
-def test_cma_mdata_matrix_prod_no_write():
-    ''' Check that we raise the correct error when we encounter a
-    kernel with only read-only CMA operators as arguments '''
-    fparser.logging.disable('CRITICAL')
-    code = CMA_MATRIX.replace('GH_WRITE', 'GH_READ', 1)
-    ast = fpapi.parse(code, ignore_comments=False)
-    name = "testkern_cma_type"
-    dkm = DynKernMetadata(ast, name=name)
-    dkm_str = str(dkm.arg_descriptors[1])
-    expected = (
-        "DynArgDescriptor03 object\n"
-        "  argument_type[0]='gh_columnwise_operator'\n"
-        "  access_descriptor[1]='gh_read'\n"
-        "  function_space_to[2]='any_space_1'\n"
-        "  function_space_from[3]='any_space_2'\n")
-    print dkm_str
-    assert expected in dkm_str
